Test: [완료] 통계 쿼리 DB에서 처리

+++ src/test/kotlin/com/group/libraryapp/service/book/BookServiceTest.kt
@@ -190,11 +190,11 @@ class BookServiceTest @Autowired constructor(
 
         //then
         assertThat(results).hasSize(2)
-        assertCount(results, BookType.COMPUTER, 3)
-        assertCount(results, BookType.SCIENCE, 1)
+        assertCount(results, BookType.COMPUTER, 3L)
+        assertCount(results, BookType.SCIENCE, 1L)
     }
 
-    private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Int) {
+    private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Long) {
         assertThat(results.first { result -> result.type == type }.count).isEqualTo(count)
     }
 }
